semana7/conexion_rs/requirements.txt
greenlet==3.0.3
numpy==1.26.4
packaging==24.1
pandas==2.0.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==1.4.53
sqlalchemy-redshift==0.8.14
typing_extensions==4.12.2
tzdata==2024.1
